requirements.txt
Django==3.0.8
django-ratelimit==3.0.0
django-extensions==3.0.2
sendgrid==6.4.3
snakemake
snakeface
django-q==1.3.4
# just for postgres
psycopg2-binary==2.8.5
django-gravatar2==1.4.4
django-taggie==1.3.0
djangorestframework==3.11.1
social-auth-app-django==4.0.0
